Fixed up video playback for playlists

+++ MediaBrowser.WindowsPhone8/Views/Playlists/AddToPlaylistView.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows.Navigation;
+using Microsoft.Phone.Controls;
 
 namespace MediaBrowser.WindowsPhone.Views.Playlists
 {
@@ -13,7 +14,7 @@ namespace MediaBrowser.WindowsPhone.Views.Playlists
         protected override void OnNavigatedFrom(NavigationEventArgs e)
         {
             base.OnNavigatedFrom(e);
-            if (e.NavigationMode != NavigationMode.Back)
+            if (e.NavigationMode != NavigationMode.Back && !(e.Content is ListPickerPage))
             {
                 NavigationService.RemoveBackEntry();
             }
